+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.70
+### is.codion.framework.model
+- EntityEditor.Modified.predicate() replaced with ObservableState based additional().
 
 ## 0.18.69
 ### is.codion.common.utilities